update enet to newer version

+++ src/net_lib/enet/types.h
@@ -1,4 +1,4 @@
-/**
+/** 
  @file  types.h
  @brief type definitions for ENet
 */
@@ -7,7 +7,7 @@
 
 typedef unsigned char enet_uint8;       /**< unsigned 8-bit type  */
 typedef unsigned short enet_uint16;     /**< unsigned 16-bit type */
-typedef size_t enet_uint32;      /**< unsigned 32-bit type */
+typedef unsigned int enet_uint32;      /**< unsigned 32-bit type */
 
 #endif /* __ENET_TYPES_H__ */
